huggingface_space/requirements.txt
gradio
torch
transformers
Pillow
accelerate
